refactor(OriasStatus): Change format of OriasStatus Enum
