+++ src/main/scala/com/particeep/api/models/partner/PartnerCompanyCreation.scala
@@ -2,7 +2,7 @@ package com.particeep.api.models.partner
 
 import play.api.libs.json.{ JsObject, Json, OFormat }
 
-import com.particeep.api.models.enums.OriasStatus.OriasStatus
+import com.particeep.api.models.enums.OriasStatus
 import com.particeep.api.models.enums.RegulatoryStatus.RegulatoryStatus
 
 case class PartnerCompanyCreation(
